Refactor video player components and implement new content structure with channels and clips
